[CALCITE-1147] Allow multiple providers for the same kind of metadata
Make it possible to extend RelMetadataQuery for a user-defined metadata
type, and create a test case illustrating.

Give an explicit error when metadata provider does not have a RelNode
handler (that is, a safety net if there are no handlers for more specific
RelNode sub-classes).

Improve how we handle ExecutionException.

(cherry picked from commit 94f8837)